requirements.txt
requests
pillow
transformers
torch
